+++ client/Packages/com.beamable.server/Editor/ManagerClient/MicroserviceManager.cs
@@ -4,7 +4,6 @@ using Beamable.Common.Api;
 using Beamable.Serialization;
 using System;
 using System.Collections.Generic;
-using UnityEngine;
 using static Beamable.Common.Constants.Features.Services;
 
 namespace Beamable.Server.Editor.ManagerClient
@@ -109,6 +108,7 @@ namespace Beamable.Server.Editor.ManagerClient
 		public string serviceName;
 		public string checksum;
 		public bool enabled;
+		public bool archived;
 		public string imageId;
 		public string templateId;
 		public string comments;
@@ -123,6 +123,7 @@ namespace Beamable.Server.Editor.ManagerClient
 		public string id;
 		public string storageType;
 		public bool enabled;
+		public bool archived;
 		public string templateId;
 		public string checksum;
 	}