+++ client/Packages/com.beamable.server/Editor/ReflectionCache/UserSystems/MicroserviceReflectionCache.cs
@@ -24,6 +24,7 @@ using UnityEditor.Callbacks;
 using UnityEngine;
 using static Beamable.Common.Constants.Features.Services;
 using static Beamable.Common.Constants.MenuItems.Assets.Orders;
+using static Beamable.Common.Constants.Features.Docker;
 using LogMessage = Beamable.Editor.UI.Model.LogMessage;
 
 namespace Beamable.Server.Editor
@@ -89,6 +90,7 @@ namespace Beamable.Server.Editor
 				_storageToBuilder.Clear();
 
 				Descriptors.Clear();
+				StorageDescriptors.Clear();
 				AllDescriptors.Clear();
 			}
 
@@ -211,6 +213,9 @@ namespace Beamable.Server.Editor
 						// TODO: XXX this is a hacky way to ignore the default microservice...
 						if (serviceAttribute.MicroserviceName.ToLower().Equals("xxxx")) continue;
 
+						if (!File.Exists(serviceAttribute.GetSourcePath()))
+							continue;
+
 						// Create descriptor
 						var hasWarning = msAttrValidationResult.Type == ReflectionCache.ValidationResultType.Warning;
 						var hasError = msAttrValidationResult.Type == ReflectionCache.ValidationResultType.Error;
@@ -280,6 +285,9 @@ namespace Beamable.Server.Editor
 						// TODO: XXX this is a hacky way to ignore the default microservice...
 						if (serviceAttribute.StorageName.ToLower().Equals("xxxx")) continue;
 
+						if (!File.Exists(serviceAttribute.SourcePath))
+							continue;
+
 						// Create descriptor
 						var hasWarning = storageObjectValResults.Type == ReflectionCache.ValidationResultType.Warning;
 						var hasError = storageObjectValResults.Type == ReflectionCache.ValidationResultType.Error;
@@ -336,6 +344,15 @@ namespace Beamable.Server.Editor
 				var nameToImageId = new Dictionary<string, string>();
 				var enabledServices = new List<string>();
 
+				
+				var availableArchitectures = await new GetBuildOutputArchitectureCommand().StartAsync();
+				
+				if (!MicroserviceConfiguration.Instance.DockerCPUArchitecture.Contains(SUPPORTED_DEPLOY_ARCHITECTURE))
+				{
+					OnDeployFailed?.Invoke(model, $"Deploy failed due to not supported Beamable Portal {MicroserviceConfiguration.Instance.DockerCPUArchitecture} architecture.");
+					return;
+				}
+				
 				foreach (var descriptor in Descriptors)
 				{
 					UpdateServiceDeployStatus(descriptor, ServicePublishState.InProgress);
@@ -353,11 +370,13 @@ namespace Beamable.Server.Editor
 
 					try
 					{
-						var buildCommand = new BuildImageCommand(descriptor,
-															 includeDebugTools: false,
-															 watch: false,
-															 pull: true);
+						var buildCommand = new BuildImageCommand(descriptor, availableArchitectures,
+						                                         includeDebugTools: false,
+						                                         watch: false,
+						                                         pull: true);
+						
 						await buildCommand.StartAsync();
+						
 					}
 					catch (Exception e)
 					{
@@ -473,6 +492,7 @@ namespace Beamable.Server.Editor
 						serviceName = sa.Name,
 						templateId = sa.TemplateId,
 						enabled = sa.Enabled,
+						archived = sa.Archived,
 						comments = sa.Comment,
 						imageId = kvp.Value,
 						dependencies = sa.Dependencies
@@ -488,6 +508,7 @@ namespace Beamable.Server.Editor
 				{
 					var sa = model.Services[uploadServiceReference.serviceName];
 					uploadServiceReference.enabled = sa.Enabled;
+					uploadServiceReference.archived = sa.Archived;
 				}
 
 				// 4- Make sure we only have each service once on the list.
@@ -510,6 +531,7 @@ namespace Beamable.Server.Editor
 						storageType = kvp.Value.Type,
 						templateId = kvp.Value.TemplateId,
 						enabled = kvp.Value.Enabled,
+						archived = kvp.Value.Archived
 					};
 				}).ToList();
 
@@ -613,6 +635,7 @@ namespace Beamable.Server.Editor
 							Comment = "",
 							Name = name,
 							Enabled = configEntry?.Enabled ?? true,
+							Archived = configEntry?.Archived ?? false,
 							TemplateId = configEntry?.TemplateId ?? "small",
 							Dependencies = serviceDependencies
 						};
@@ -638,6 +661,7 @@ namespace Beamable.Server.Editor
 							Name = name,
 							Type = configEntry?.StorageType ?? "mongov1",
 							Enabled = configEntry?.Enabled ?? true,
+							Archived = configEntry?.Archived ?? false,
 							TemplateId = configEntry?.TemplateId ?? "small",
 						};
 					}).ToList();