+++ microservice/microserviceTests/microservice/Util/LoggingUtil.cs
@@ -2,23 +2,39 @@ using Beamable.Common;
 using Beamable.Server;
 using Core.Server.Common;
 using Serilog;
+using Serilog.Events;
 using Serilog.Formatting.Raw;
+using Serilog.Sinks.TestCorrelator;
 using UnityEngine;
 
 namespace microserviceTests.microservice.Util
 {
     public class LoggingUtil
     {
-        public static void Init()
+        public static void Init(LogEventLevel logLevel=LogEventLevel.Warning)
         {
             BeamableLogProvider.Provider = new BeamableSerilogProvider();
             Debug.Instance = new MicroserviceDebug();
             // https://github.com/serilog/serilog/wiki/Configuration-Basics
             Log.Logger = new LoggerConfiguration()
-               .MinimumLevel.Warning()
+               .MinimumLevel.Is(logLevel)
                .WriteTo.Console(new RawFormatter())
                .CreateLogger();
             BeamableSerilogProvider.LogContext.Value = Log.Logger;
+
+        }
+
+        public static void InitTestCorrelator(LogEventLevel logLevel=LogEventLevel.Warning)
+        {
+	        TestCorrelator.CreateContext();
+	        BeamableLogProvider.Provider = new BeamableSerilogProvider();
+	        Debug.Instance = new MicroserviceDebug();
+	        // https://github.com/serilog/serilog/wiki/Configuration-Basics
+	        Log.Logger = new LoggerConfiguration()
+		        .MinimumLevel.Is(logLevel)
+		        .WriteTo.TestCorrelator()
+		        .CreateLogger();
+	        BeamableSerilogProvider.LogContext.Value = Log.Logger;
 
         }
     }