+++ microservice/microserviceTests/microservice/dbmicroservice/BeamableMicroServiceTests/StartTests.cs
@@ -14,6 +14,11 @@ using Beamable.Server.Content;
 using microserviceTests.microservice.Util;
 using Newtonsoft.Json.Linq;
 using NUnit.Framework;
+using Serilog.Events;
+using Serilog.Sinks.TestCorrelator;
+using System;
+using System.Diagnostics;
+using System.Linq;
 using ClientRequest = Beamable.Microservice.Tests.Socket.ClientRequest;
 
 namespace microserviceTests.microservice.dbmicroservice.BeamableMicroServiceTests
@@ -25,7 +30,8 @@ namespace microserviceTests.microservice.dbmicroservice.BeamableMicroServiceTest
         [TearDown]
         public void ResetContentInstance()
         {
-            ContentApi.Instance = new Promise<IContentApi>();
+	        ContentApi.Instance = new Promise<IContentApi>();
+	        BeamableMicroService._contentService = null;
         }
 
         [Test]
@@ -381,7 +387,7 @@ namespace microserviceTests.microservice.dbmicroservice.BeamableMicroServiceTest
         [NonParallelizable]
         public async Task HandleSimple_FailGetUserViaAccessToken()
         {
-            LoggingUtil.Init();
+            LoggingUtil.InitTestCorrelator();
             TestSocket testSocket = null;
             const int testCount = 3000;
             var contentResolver = new TestContentResolver(async uri => "{}");
@@ -412,6 +418,14 @@ namespace microserviceTests.microservice.dbmicroservice.BeamableMicroServiceTest
 
             await Task.WhenAll(tasks);
             await Task.Delay(10);
+
+            var notImplementedLogs = TestCorrelator.GetLogEventsFromCurrentContext()
+	            .Where(l => l.Level == LogEventLevel.Error)
+	            .Select(l => l.RenderMessage())
+	            .Where(l => l.Contains("This version of GetUser is not supported in the Microservice environment"))
+	            .ToList();
+            Assert.AreEqual(testCount, notImplementedLogs.Count);
+
             // simulate shutdown event...
             await ms.OnShutdown(this, null);
             //Assert.IsTrue(testSocket.AllMocksCalled());
@@ -842,7 +856,7 @@ namespace microserviceTests.microservice.dbmicroservice.BeamableMicroServiceTest
         [NonParallelizable]
         public async Task HandleConnectionDrop_WhileMessageInFlight()
         {
-            LoggingUtil.Init();
+            LoggingUtil.Init(LogEventLevel.Verbose);
             TestSocket testSocket = null;
             var contentResolver = new TestContentResolver();
             var connectionIndex = 0;
@@ -859,9 +873,19 @@ namespace microserviceTests.microservice.dbmicroservice.BeamableMicroServiceTest
                 {
                     socket
                     .WithName("second")
-                    .AddStandardMessageHandlers(4)
+                    .AddAuthMessageHandlersWithDelay(1000, 1000, 4, 1)
+                    .AddProviderMessageHandlers(5)
                     .AddMessageHandler(
-                       MessageMatcher.WithReqId(1).WithStatus(200).WithPayload<int>(n => n == 500),
+	                    MessageMatcher
+		                    .WithRouteContains("basic/accounts")
+		                    .WithBody<dynamic>(d => d.gamerTag == 123),
+	                    MessageResponder.SuccessWithDelay(25, new User
+	                    {
+		                    email = "fakeEmail"
+	                    }),
+	                    MessageFrequency.OnlyOnce())
+                    .AddMessageHandler(
+                       MessageMatcher.WithReqId(1).WithStatus(200).WithPayload<string>(n => n == "fakeEmail"),
                        MessageResponder.NoResponse(),
                        MessageFrequency.OnlyOnce()
                     );
@@ -873,14 +897,14 @@ namespace microserviceTests.microservice.dbmicroservice.BeamableMicroServiceTest
             await ms.Start<SimpleMicroservice>(new TestArgs());
             Assert.IsTrue(ms.HasInitialized);
 
-            testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "Delay", 1, 1, 500));
+            testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "DelayThenGetEmail", 1, 1, 500, 123));
 
             // simulate connection drop.
             Assert.IsTrue(testSocket.AllMocksCalled());
             testSocket.Fault();
 
             // wait longer than the message's delay..
-            await Task.Delay(550);
+            await Task.Delay(1550);
 
             // simulate shutdown event...
             await ms.OnShutdown(this, null);
@@ -1199,6 +1223,7 @@ namespace microserviceTests.microservice.dbmicroservice.BeamableMicroServiceTest
 
             testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "GetUserEmail", 1, 1, dbid));
 
+            await Task.Delay(1000); // the auth cycle will take some time to figure itself out, and if we trigger a shutdown too soon, the request id of the shutdown will get bungled with the re-auth flow.
             await ms.OnShutdown(this, null);
             Assert.IsTrue(testSocket.AllMocksCalled());
         }
@@ -1217,9 +1242,19 @@ namespace microserviceTests.microservice.dbmicroservice.BeamableMicroServiceTest
             var authDelay = 100;
             const int failureCount = 2000;
             var sent = false;
+            var noncePromise = new Promise();
+            TestSocketResponseGeneratorAsync nonceSuccess = async res =>
+            {
+	            await noncePromise;
+	            return res.Succeed(new MicroserviceNonceResponse { nonce = "testnonce" });
+            };
+            int successCount = 0;
+            int authFailCount = 0;
+
             var ms = new BeamableMicroService(new TestSocketProvider(socket =>
             {
                 testSocket = socket;
+
                 socket
                 .WithName("first")
                 .AddStandardMessageHandlers()
@@ -1227,31 +1262,41 @@ namespace microserviceTests.microservice.dbmicroservice.BeamableMicroServiceTest
                    MessageMatcher
                       .WithRouteContains("basic/accounts")
                       .WithBody<dynamic>(d => d.gamerTag == dbid),
-                   MessageResponder.AuthFailure(),
-                   MessageFrequency.Exactly(failureCount))
+                   MessageResponder.Custom(res =>
+                   {
+	                   if (noncePromise.IsCompleted)
+	                   {
+		                   Interlocked.Increment(ref successCount);
+		                   return res.Succeed(new User { email = fakeEmail });
+	                   }
+
+	                   Interlocked.Increment(ref authFailCount);
+	                   return res.AuthFailure();
+                   }),
+                   MessageFrequency.AtLeast(1),
+                   "original failure"
+                   )
                 .AddMessageHandler(
                    MessageMatcher.WithRouteContains("nonce"),
-                   MessageResponder.SuccessAfterCondition(() => sent, new MicroserviceNonceResponse { nonce = "testnonce" }),
-                   MessageFrequency.OnlyOnce()
+                   MessageResponder.CustomAsync(nonceSuccess),
+                   MessageFrequency.OnlyOnce(), "re-nonce"
                 )
                 .AddMessageHandler(
                    MessageMatcher.WithRouteContains("auth"),
-                   MessageResponder.SuccessWithDelay(authDelay, new MicroserviceAuthResponse { result = "ok" }),
-                   MessageFrequency.OnlyOnce()
+                   MessageResponder.SuccessWithDelay(authDelay, ()=>
+                   {
+	                   testSocket.IsAuthenticated = true;
+	                   return new MicroserviceAuthResponse { result = "ok" };
+                   }),
+                   MessageFrequency.OnlyOnce(), "re-auth"
                 )
-                .AddMessageHandler(
-                   MessageMatcher
-                      .WithRouteContains("basic/accounts")
-                      .WithBody<dynamic>(d => d.gamerTag == dbid),
-                   MessageResponder.Success(new User { email = fakeEmail }),
-                   MessageFrequency.Exactly(failureCount)) // allow 2 retries...
-
                 .AddMessageHandler(
                    MessageMatcher
                       .WithPositiveReqId()
                       .WithStatus(200).WithPayload<string>(n => n == fakeEmail),
                    MessageResponder.NoResponse(),
-                   MessageFrequency.Exactly(failureCount)
+                   MessageFrequency.Exactly(failureCount),
+                   "success-after-failure"
                 )
                 ;
 
@@ -1264,20 +1309,23 @@ namespace microserviceTests.microservice.dbmicroservice.BeamableMicroServiceTest
             var tasks = new List<Task>();
             for (var i = 0; i < failureCount; i++)
             {
-                tasks.Add(Task.Run(() =>
+	            var index = i + 1;
+	            tasks.Add(Task.Run(() =>
                 {
-                    testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "GetUserEmail", i + 1, 1, dbid));
+                    testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "GetUserEmail", index, 1, dbid));
                 }));
             }
-
             await Task.Delay(nonceDelay);
-            sent = true;
+            noncePromise.CompleteSuccess();
 
-            await Task.WhenAll(tasks);
-            // testSocket.SendToClient(ClientRequest.ClientCallable("micro_sample", "GetUserEmail", 2, 0, dbid));
+            var waitingTask = Task.WhenAll(tasks);
+            await waitingTask;
 
+            await Task.Delay(1000);
             await ms.OnShutdown(this, null);
             Assert.IsTrue(testSocket.AllMocksCalled());
+            Assert.IsTrue(authFailCount <= failureCount); // it cannot be the case that we failed more messages for auth failure than we had original requests.
+            Assert.AreEqual(failureCount, successCount); // we need to make sure that the success message came back for each failure, eventually.
         }
 
     }