+++ client/Packages/com.beamable.server/CHANGELOG.md
@@ -4,6 +4,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- User can specify Microservices build CPU architecture. 
+- `RemovePlayerEntry` for leaderboards API which allows to remove given player from the leaderboard
+- Added microservice archive/unarchive feature.
+
+## [1.2.6]
+### Added
+- `RemovePlayerEntry` for leaderboards API which allows to remove given player from the leaderboard
+
+### Fixed
+- Microservices may be built from either ARM or x86 based computers and uploaded to Beamable.
+
+## [1.2.5]
+### Fixed
+- Failed promises no longer log exception info after an exception handler is registered on the same execution cycle.
+- "Connection is closed" log exception no longer prints incorrectly.
+- Requests no longer attempt to send while authorization process is happening.
+
 ## [1.2.4]
 ### Added
 - Microservices now support private declarations of `Callable` methods.